+++ yave/graphics/device/DeviceProperties.h
@@ -27,10 +27,10 @@ SOFTWARE.
 namespace yave {
 
 struct DeviceProperties {
-    u32 non_coherent_atom_size;
+    u64 non_coherent_atom_size;
     u32 max_uniform_buffer_size;
-    u32 uniform_buffer_alignment;
-    u32 storage_buffer_alignment;
+    u64 uniform_buffer_alignment;
+    u64 storage_buffer_alignment;
 
     u32 max_memory_allocations;
